+++ test/messages/MessagesBuilderStream.ts
@@ -6,12 +6,12 @@ import { MessagesBuilder } from '../../src/messages/MessagesBuilder.js'
 export class MessagesBuilderStream extends Transform {
   private readonly builder = new MessagesBuilder()
 
-  constructor() {
+  constructor(private readonly errorHandler: (err: Error) => void = () => undefined) {
     super({ objectMode: true })
   }
 
   _transform(envelope: Envelope, _: BufferEncoding, callback: TransformCallback) {
-    this.builder.processEnvelope(envelope)
+    this.builder.processEnvelope(envelope, this.errorHandler)
     callback()
   }