+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add optional error handler to `MessageBuilder#processMessage`
 
 ## [0.8.0] - 2021-10-21
 ### Added